Fix #18: Make xderiving macro not rely on Scalaz syntax
The xderiving macro relied on the InvariantFunctor syntax from Scalaz to
work. This indirection required the scalaz syntax in scope in order to
work and threw confusing errors if it wasn't in scope. Additionally, the
syntax doesn't always get picked up.

This commit removes the syntax requirement while still maintaining the
raw-xmap feature.